Enable allocation hoisting out of loops (#43057)
