+++ src/llvm-julia-licm.cpp
@@ -8,9 +8,12 @@
 #include <llvm/IR/Dominators.h>
 #include <llvm/IR/LegacyPassManager.h>
 #include <llvm/Transforms/Utils/LoopUtils.h>
+#include <llvm/Analysis/ValueTracking.h>
 
 #include "llvm-pass-helpers.h"
 #include "julia.h"
+#include "llvm-alloc-helpers.h"
+#include "codegen_shared.h"
 
 #define DEBUG_TYPE "julia-licm"
 
@@ -37,11 +40,13 @@ struct JuliaLICMPass : public LoopPass, public JuliaPassContext {
         if (!preheader)
             return false;
         BasicBlock *header = L->getHeader();
+        const llvm::DataLayout &DL = header->getModule()->getDataLayout();
         initFunctions(*header->getModule());
         // Also require `gc_preserve_begin_func` whereas
         // `gc_preserve_end_func` is optional since the input to
         // `gc_preserve_end_func` must be from `gc_preserve_begin_func`.
-        if (!gc_preserve_begin_func)
+        // We also hoist write barriers here, so we don't exit if write_barrier_func exists
+        if (!gc_preserve_begin_func && !write_barrier_func && !alloc_obj_func)
             return false;
         auto LI = &getAnalysis<LoopInfoWrapperPass>().getLoopInfo();
         auto DT = &getAnalysis<DominatorTreeWrapperPass>().getDomTree();
@@ -110,6 +115,44 @@ struct JuliaLICMPass : public LoopPass, public JuliaPassContext {
                     for (unsigned i = 1; i < exit_pts.size(); i++) {
                         // Clone exit
                         CallInst::Create(call, {}, exit_pts[i]);
+                    }
+                }
+                else if (callee == write_barrier_func) {
+                    bool valid = true;
+                    for (std::size_t i = 0; i < call->getNumArgOperands(); i++) {
+                        if (!L->makeLoopInvariant(call->getArgOperand(i), changed)) {
+                            valid = false;
+                            break;
+                        }
+                    }
+                    if (valid) {
+                        call->moveBefore(preheader->getTerminator());
+                        changed = true;
+                    }
+                }
+                else if (callee == alloc_obj_func) {
+                    jl_alloc::AllocUseInfo use_info;
+                    jl_alloc::CheckInst::Stack check_stack;
+                    jl_alloc::EscapeAnalysisRequiredArgs required{use_info, check_stack, *this, DL};
+                    jl_alloc::runEscapeAnalysis(call, required, jl_alloc::EscapeAnalysisOptionalArgs().with_valid_set(&L->getBlocksSet()));
+                    if (use_info.escaped || use_info.addrescaped) {
+                        continue;
+                    }
+                    bool valid = true;
+                    for (std::size_t i = 0; i < call->getNumArgOperands(); i++) {
+                        if (!L->makeLoopInvariant(call->getArgOperand(i), changed)) {
+                            valid = false;
+                            break;
+                        }
+                    }
+                    if (use_info.refstore) {
+                        // We need to add write barriers to any stores
+                        // that may start crossing generations
+                        continue;
+                    }
+                    if (valid) {
+                        call->moveBefore(preheader->getTerminator());
+                        changed = true;
                     }
                 }
             }
